requirements.txt
opencv-python
mediapipe
numpy
scikit-learn
tensorflow
keras
pyserial
